Refactor main slider with ES6 destructuring, now slider is main class, and
