+++ BringItUp/src/js/main.js
@@ -1,8 +1,8 @@
-import Slider from './modules/slider.js';
+import MainSlider from './modules/slider/sliderMain';
 import VideoPlayer from './modules/playVideo.js';
 
 window.addEventListener('DOMContentLoaded', () => {
-    const slider = new Slider('.page', '.slide', '.next', '.sidecontrol .reset');
+    const slider = new MainSlider({page: '.page', slides: '.slide', btns: '.next', resetSlidesSelector: '.sidecontrol .reset'});
     slider.render();
 
     const player = new VideoPlayer('.showup .play', '.overlay');
